Backend PlacePrep/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
sqlalchemy==2.0.29
aiosqlite==0.20.0
redis==5.0.4
groq==0.9.0
SpeechRecognition==3.10.4
pydub==0.25.1
sentence-transformers==2.7.0
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
PyPDF2==3.0.1
PyJWT==2.8.0